[flang] Add check for constraints on sync-stat-lists
In the CoarrayChecker, add checks for the constraints C1172 and
C1173, which constrain sync-stat-list. There are sync-stat-lists
in critical-stmt, sync-all-stmt, sync-images-stmt,
sync-memory-stmt, sync-team-stmt, event-post-stmt, unlock-stmt,
change-team-stmt, and end-change-team-stmt. Also add a check for
the constraint C1174 in sync-images-stmt, for the type of the
event-variable in event-post-stmt and event-wait-stmt, and check
the constraints C1177 and C1178 in event-wait-stmt. Update
semantics tests for these stmts.

Reviewed By:

Differential Revision: