[flang][OpenMP][DoConcurrent] Simplify loop-nest detection logic
With https://github.com/llvm/llvm-project/pull/114020, do-concurrent loop-nests are more conforment to the spec and easier to detect. All we need to do is to check that the only operations inside `loop A` which perfectly wraps `loop B` are:

  * the operations needed to update `loop A`'s iteration variable and
  * `loop B` itself.

This PR simlifies the pass a bit using the above logic and replaces https://github.com/ROCm/llvm-project/pull/127.